Task2/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
tenacity
pydantic